% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{addComponentInput}
\alias{addComponentInput}
\title{Adds an input to a sim component.}
\usage{
addComponentInput(
  sol,
  componentid,
  id,
  source = NULL,
  value = NULL,
  datatype = NULL,
  unit = NULL,
  description = NULL
)
}
\arguments{
\item{sol}{solution object}

\item{componentid}{id of sim component}

\item{id}{id of variable}

\item{source}{source of variable}

\item{value}{value of variable (is used only if source is not given or NULL)}

\item{datatype}{datatype (optional)}

\item{unit}{unit (optional)}

\item{description}{short description (optional)}
}
\value{
modified solution object
}
\description{
Notice: One can only add inputs that are defined by the sim component. If source
is not given, then the parameter value is used.
}
