% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{addCSVOutput}
\alias{addCSVOutput}
\title{Adds a new CSV output to solution.}
\usage{
addCSVOutput(
  sol,
  filename,
  outputid,
  frequence = "DAILY",
  rule = NULL,
  resetrule = NULL,
  cachesize = 10,
  divider = ","
)
}
\arguments{
\item{sol}{solution object}

\item{filename}{name of the output file}

\item{outputid}{id of new output}

\item{frequence}{one of DAILY, YEARLY, BOOLEAN, COMPLEX}

\item{rule}{optional rule, when frequence is BOOLEAN or COMPLEX}

\item{resetrule}{optional resetrule, when frequence is BOOLEAN or COMPLEX}

\item{cachesize}{optional cachesize}

\item{divider}{character that is used as a divider for the csv file}
}
\value{
modified solution object
}
\description{
Any output with the same id will be removed. The added output has no output
variables. One has to add new variables via \code{addOutputVariable}.
}
