% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{addMemoryOutput}
\alias{addMemoryOutput}
\title{Adds a new memory output to solution.}
\usage{
addMemoryOutput(
  sol,
  outputid,
  frequence = "DAILY",
  rule = NULL,
  resetrule = NULL,
  cachesize = 10
)
}
\arguments{
\item{sol}{solution object}

\item{outputid}{id of new output}

\item{frequence}{one of DAILY, YEARLY, BOOLEAN, COMPLEX}

\item{rule}{optional rule, when frequence is BOOLEAN or COMPLEX}

\item{resetrule}{optional resetrule, when frequence is BOOLEAN or COMPLEX}

\item{cachesize}{optional cachesize}
}
\value{
modified solution object
}
\description{
Any output with the same id will be removed. The added output has no output
variables. One has to add new variables via \code{addOutputVariable}.
}
