% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{addOutputVariable}
\alias{addOutputVariable}
\title{Adds output variable to an existing output}
\usage{
addOutputVariable(
  sol,
  outputid,
  id,
  rule,
  datatype,
  mode = NULL,
  unit = NULL,
  description = NULL
)
}
\arguments{
\item{sol}{solution object}

\item{outputid}{id of output where the variable should be added}

\item{id}{name of the new variable}

\item{rule}{rule for the variable}

\item{datatype}{of the variable}

\item{mode}{one of FIRST, LAST, AVG, SUM (optional)}

\item{unit}{unit of the variable (optional)}

\item{description}{short description (optional)}
}
\value{
modified solution object
}
\description{
Adds output variable to an existing output
}
