% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{addTimingSimComponent}
\alias{addTimingSimComponent}
\title{Adds elements for timing the execution time of sim components (and transformers)}
\usage{
addTimingSimComponent(
  sol,
  filename = NULL,
  componentlist = NULL,
  interfaceid = "automatic_timing_interface",
  outputid = "automatic_timing_output",
  simcomponentid = "AutomaticTiming"
)
}
\arguments{
\item{sol}{solution object}

\item{filename}{optional filename to write the timing information to csv file}

\item{componentlist}{optional list of component ids (if empty, all components will be timed)}

\item{interfaceid}{id for the interface (optional)}

\item{outputid}{id for the output (optional)}

\item{simcomponentid}{id for the timing simcomponent (optional)}
}
\description{
Adds elements for timing the execution time of sim components (and transformers)
}
