% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{addUserVariable}
\alias{addUserVariable}
\title{Adds an user variable to the \code{variables} section}
\usage{
addUserVariable(
  sol,
  id,
  value,
  datatype,
  unit = NULL,
  description = NULL,
  rule = NULL
)
}
\arguments{
\item{sol}{solution object}

\item{id}{id of the variable}

\item{value}{value of the variable}

\item{datatype}{of the variable}

\item{unit}{unit (optional)}

\item{description}{short description (optional)}

\item{rule}{optional rule}
}
\value{
modified solution object
}
\description{
Adds an user variable to the \code{variables} section
}
