% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{addXMLParameter}
\alias{addXMLParameter}
\title{Adds a parameter to a specific parameterset}
\usage{
addXMLParameter(paramdata, key, id, value = "", unit = "", description = NULL)
}
\arguments{
\item{paramdata}{parameterdata (xml2 object)}

\item{key}{named vector \code{c(idname="value")} to identify the dataset, the name is the id of the parameter and the value has to match the parameter's value}

\item{id}{name of the parameter which should added}

\item{value}{value to set (numeric or character vector)}

\item{unit}{unit of the parameter}

\item{description}{description of the parameter}
}
\value{
modified parameter set (xml2 object)
}
\description{
When the parameter already exists in the parameter set,
it is removed and the new parameter is added at the end.
}
\examples{
params <- readXMLParameterFile(system.file("input", "crop.xml", package="simplaceUtil"))
getXMLParameter(params, c(CROPNAME="soy bean"), "ExtraTableYZZ")
params_new <- addXMLParameter(params, c(CROPNAME="soy bean"), "ExtraTableYZZ", c(1,2,4))
getXMLParameter(params_new, c(CROPNAME="soy bean"), "ExtraTableYZZ")
}
