% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{changeAllOutputTypesToMemory}
\alias{changeAllOutputTypesToMemory}
\title{Changes all non-MEMORY outputs from solution to MEMORY outputs.}
\usage{
changeAllOutputTypesToMemory(sol)
}
\arguments{
\item{sol}{solution object}
}
\value{
modified solution object
}
\description{
When running large amout of runs (e.g. calibration) it's recommended to avoid
to write outputs on disk.
}
