% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{changeInterfaceFile}
\alias{changeInterfaceFile}
\title{Changes all non-MEMORY outputs from solution to MEMORY outputs.}
\usage{
changeInterfaceFile(sol, id, filename, keeppath = FALSE)
}
\arguments{
\item{sol}{solution object}

\item{id}{id of interface}

\item{filename}{new filename}

\item{keeppath}{if set to true,  original path is kept and only filename is modified}
}
\value{
modified solution object
}
\description{
When running large amout of runs (e.g. calibration) it's recommended to avoid
to write outputs on disk.
}
