% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{changeOutputType}
\alias{changeOutputType}
\title{Changes the type of an output from CSV to MEMORY or vice versa}
\usage{
changeOutputType(sol, outputid, type, filename = NULL, divider = ",")
}
\arguments{
\item{sol}{solution object}

\item{outputid}{id of output to change}

\item{type}{one of 'CSV' or 'MEMORY'}

\item{filename}{name of the output file if type changes to 'CSV'}

\item{divider}{character that is used as a divider for the csv file}
}
\value{
modified solution object
}
\description{
Changes the type of an output from CSV to MEMORY or vice versa
}
