% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_generation.R
\name{createCodeStubsForSimVariables}
\alias{createCodeStubsForSimVariables}
\title{Create java and xml code stubs from a csv file}
\usage{
createCodeStubsForSimVariables(
  variables,
  component = "MyComponent",
  outfolder = NULL,
  ...
)
}
\arguments{
\item{variables}{dataframe or path to csv file with variables}

\item{component}{name of SimComonent}

\item{outfolder}{optional, if given code is written the folder's files}

\item{...}{parameters passed to read.table (e.g. sep, dec etc.)}
}
\value{
list of code snippets
}
\description{
The function creates Java code stubs for defining the fields, adding
variables and initialising values.
It creates XML subs for parameter file, as well as resource definition for
parameter and input data, inputs for simcomponent and outputs.
}
\details{
The dataframe / CSV file should have columns
\itemize{
\item contenttype (one of: constant, input, state, rate, out)
\item id
\item description
\item datatype (Simplace Datatype or corresponding Java Datatype)
\item unit
\item min
\item max
\item default
}
}
