% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{createCsvResourceStubs}
\alias{createCsvResourceStubs}
\title{Creates XML stubs for interface and resource section from CSV file structure}
\usage{
createCsvResourceStubs(
  filename,
  id,
  sep = ",",
  keyvals = NULL,
  arraycolumns = NULL,
  frequence = "DAILY",
  rule = NULL,
  data = NULL
)
}
\arguments{
\item{filename}{name of the CSV file}

\item{id}{id for the resource}

\item{sep}{separator for CSV file}

\item{keyvals}{named vector of column indices or names that act as key column}

\item{arraycolumns}{vector of column indices or names that are arrays}

\item{frequence}{frequence attribute for the resource}

\item{rule}{rule attribute for the resource}

\item{data}{optional data.frame (otherwise data will be loaded from file)}
}
\value{
list of two strings ('interface' and 'resource')
}
\description{
Creates XML stubs for interface and resource section from CSV file structure
}
\examples{
stubs <- createCsvResourceStubs(
  filename = system.file("input","weather.csv", package="simplaceUtil"),
  id = "weather",
  sep =",",
  keyvals = c("CURRENT.DATE" = "Date")
)
cat(stubs$interface)
cat(stubs$resource)

stubs <- createCsvResourceStubs(
  filename = system.file("input","soil.csv", package="simplaceUtil"),
  id = "soil",
  sep =";",
  keyvals = c("vSoilType" = "soiltype"),
  arraycolumns = 6:20
)
cat(stubs$resource)

}
