% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{createXmlResourceStubs}
\alias{createXmlResourceStubs}
\title{Creates XML stubs for interface and resource section from XML file structure}
\usage{
createXmlResourceStubs(
  filename,
  id,
  keyvals = NULL,
  frequence = "DAILY",
  rule = NULL,
  xmlnode = NULL
)
}
\arguments{
\item{filename}{name of the xml file}

\item{id}{id for the resource}

\item{keyvals}{named vector of column indices or names that act as key column}

\item{frequence}{frequence attribute for the resource}

\item{rule}{rule attribute for the resource}

\item{xmlnode}{optional xml_node (otherwise xml will be read from filename)}
}
\value{
list of two strings ('interface' and 'resource')
}
\description{
Creates XML stubs for interface and resource section from XML file structure
}
\examples{
stubs <- createXmlResourceStubs(
  filename = system.file("input","crop.xml",package="simplaceUtil"),
  id="soil",
  keyvals = c("vSoilType"=1)
)
cat(stubs$resource)
}
