% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website_documentation.R
\name{fetchDescriptionFromWebsite}
\alias{fetchDescriptionFromWebsite}
\title{Fetches the Description (as HTML source code) of a SimComponent from Simplace Website}
\usage{
fetchDescriptionFromWebsite(class, version = "current")
}
\arguments{
\item{class}{class name of the SimComponent}

\item{version}{Use \code{current}, \code{trunk} or numeric version \verb{#.#}}
}
\value{
string with the description text (including HTML tags)
}
\description{
Fetches the Description (as HTML source code) of a SimComponent from Simplace Website
}
