% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website_documentation.R
\name{fetchSimComponentlistFromWebsite}
\alias{fetchSimComponentlistFromWebsite}
\title{Fetches the list of SimComponents from the Simplace Website}
\usage{
fetchSimComponentlistFromWebsite(version = "current")
}
\arguments{
\item{version}{Use \code{current}, \code{trunk} or numeric version \verb{#.#}}
}
\value{
character vector with the class names of all SimComponents
}
\description{
Fetches the list of SimComponents from the Simplace Website
}
