% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website_documentation.R
\name{fetchSimVariablesFromWebsite}
\alias{fetchSimVariablesFromWebsite}
\title{Fetches the SimVariables table for a SimComponent from Simplace Website}
\usage{
fetchSimVariablesFromWebsite(class, version = "current")
}
\arguments{
\item{class}{class name of the SimComponent}

\item{version}{Use \code{current}, \code{trunk} or numeric version \verb{#.#}}
}
\value{
a data.frame with the SimVariables
}
\description{
Fetches the SimVariables table for a SimComponent from Simplace Website
}
