% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sol_and_proj_formatting.R
\name{formatSolutionFile}
\alias{formatSolutionFile}
\title{Formats / prettyprints a solution file}
\usage{
formatSolutionFile(
  file,
  outfile = file,
  version = "5.2",
  stripcomments = FALSE,
  dropattributes = c()
)
}
\arguments{
\item{file}{filename of solution or project}

\item{outfile}{filename the reformatted document is written to}

\item{version}{version of the solution}

\item{stripcomments}{if TRUE, then strip xml comments}

\item{dropattributes}{vector of attribute names that will be dropped globally.}
}
\value{
no return value, called for the side effect to write a file to disk
}
\description{
Formats / prettyprints a solution file
}
\examples{
\dontrun{
original <- system.file("solution","Yield.sol.xml",package="simplaceUtil")
formatted <- tempfile(fileext = ".sol.xml") 
formatSolutionFile(original, formatted, stripcomments = TRUE)
}
}
