% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionToGraphFunctions.R
\name{getComponentsFromFile}
\alias{getComponentsFromFile}
\title{Get component and links dataframe from solution file}
\usage{
getComponentsFromFile(file)
}
\arguments{
\item{file}{solution}
}
\description{
Get component and links dataframe from solution file
}
