% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getElementsFromSolutionFile}
\alias{getElementsFromSolutionFile}
\title{Get component and links dataframe from solution file}
\usage{
getElementsFromSolutionFile(file)
}
\arguments{
\item{file}{solution}
}
\description{
Get component and links dataframe from solution file
}
