% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionToGraphFunctions.R
\name{getLinkingFromComponent}
\alias{getLinkingFromComponent}
\title{Get components that are linked from given component}
\usage{
getLinkingFromComponent(graph, name, distance = 50, linkage = "allsteps")
}
\arguments{
\item{graph}{graph}

\item{name}{name of component}

\item{distance}{maximum distance from given component}

\item{linkage}{all steps, same step or previous step}
}
\description{
Get components that are linked from given component
}
