% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_to_graph.R
\name{getLinkingToComponent}
\alias{getLinkingToComponent}
\title{Get components that link to given component}
\usage{
getLinkingToComponent(graph, name, distance = 50, linkage = "allsteps")
}
\arguments{
\item{graph}{graph}

\item{name}{name of component}

\item{distance}{maximum distance from given component}

\item{linkage}{all steps, same step or previous step}
}
\description{
Get components that link to given component
}
