% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getLinks}
\alias{getLinks}
\title{Get the dataframe of links between components}
\usage{
getLinks(x, df)
}
\arguments{
\item{x}{xml object (solution)}

\item{df}{data.frame of components}
}
\value{
data.frame with the linked variables between components
}
\description{
Get the dataframe of links between components
}
