% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getMemoryOutputIds}
\alias{getMemoryOutputIds}
\title{Get ids of memory outputs}
\usage{
getMemoryOutputIds(comp)
}
\arguments{
\item{comp}{components dataframe}
}
\value{
character vector with the memory output ids
}
\description{
Get ids of memory outputs
}
