% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getMetadataForFile}
\alias{getMetadataForFile}
\title{Get metadata from file}
\usage{
getMetadataForFile(file, workdir)
}
\arguments{
\item{file}{filename (full path)}

\item{workdir}{working directory path}
}
\description{
Get metadata from file
}
\keyword{internal}
