% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_to_graph.R
\name{getNeighborhood}
\alias{getNeighborhood}
\title{Get the simcomponents in the neighborhood of given components}
\usage{
getNeighborhood(
  graph,
  names,
  distance = 1,
  linkage = "allsteps",
  type = "all",
  set_op = "union"
)
}
\arguments{
\item{graph}{graph object of class \code{dgr_graph}}

\item{names}{names (id attributes) of the selected sim components}

\item{distance}{maximum number of steps from given component}

\item{linkage}{all steps, same step or previous step}

\item{type}{type of component}

\item{set_op}{"union" - all neighbours of selected components or
"intersect" - common neighbours of all selectected components}
}
\value{
graph object of class \code{dgr_graph}
}
\description{
Get the simcomponents in the neighborhood of given components
}
