% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getOutputFilenames}
\alias{getOutputFilenames}
\title{Get filenames of file outputs}
\usage{
getOutputFilenames(comp, variables, additional = NULL)
}
\arguments{
\item{comp}{components dataframe}

\item{variables}{variables dataframe}

\item{additional}{additional variables as named vector c("var1"="value1", ...), useful for directory placeholder}
}
\value{
named character vector with the filenames
}
\description{
Get filenames of file outputs
}
