% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getResourceVariables}
\alias{getResourceVariables}
\title{Get variables for a specific resource or transformer}
\usage{
getResourceVariables(x, id, compactkey = FALSE)
}
\arguments{
\item{x}{xml object (solution)}

\item{id}{id of the resource or transformer}

\item{compactkey}{if TRUE, then id and key column are merged}
}
\value{
a datatframe with all the variables of a resource
}
\description{
Get variables for a specific resource or transformer
}
