% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{getSolutionFromFile}
\alias{getSolutionFromFile}
\title{Reads a solution from file}
\usage{
getSolutionFromFile(file)
}
\arguments{
\item{file}{filename of the solution}
}
\value{
parsed solution as xml_document
}
\description{
Reads a solution from file
}
