% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{getSolutionFromText}
\alias{getSolutionFromText}
\title{Reads a solution from text}
\usage{
getSolutionFromText(text)
}
\arguments{
\item{text}{string with xml markup}
}
\value{
parsed solution as xml_document
}
\description{
Reads a solution from text
}
