% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{getTextFromSolution}
\alias{getTextFromSolution}
\title{Converts a solution to text}
\usage{
getTextFromSolution(sol)
}
\arguments{
\item{sol}{solution object (xml_document)}
}
\value{
text with xml code
}
\description{
Converts a solution to text
}
