% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{getXMLParameter}
\alias{getXMLParameter}
\title{Get a parameter value for a specific parameterset}
\usage{
getXMLParameter(paramdata, key, id)
}
\arguments{
\item{paramdata}{parameterdata (xml2 object)}

\item{key}{named vector \code{c(idname="value")} to identify the dataset, the name is the id of the parameter and the value has to match the parameter's value}

\item{id}{name of the parameter of which the value should be retrieved}
}
\value{
the value of the parameter
}
\description{
Get a parameter value for a specific parameterset
}
\examples{

params <- readXMLParameterFile(system.file("input", "crop.xml", package="simplaceUtil"))
getXMLParameterIDs(params, "CROPNAME")
getXMLParameter(params, c(CROPNAME="soy bean"), "TSUM1")
getXMLParameter(params, c(CROPNAME="faba bean"), "TSUM1")

}
