% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{getXMLParameterIDs}
\alias{getXMLParameterIDs}
\title{Get IDs of all parametersets in the parameterfile}
\usage{
getXMLParameterIDs(paramdata, keyid)
}
\arguments{
\item{paramdata}{parameterdata (xml2 object)}

\item{keyid}{id of the parameter that identifies the parameterset}
}
\value{
character vector with all IDs
}
\description{
Get IDs of all parametersets in the parameterfile
}
\examples{

params <- readXMLParameterFile(system.file("input", "crop.xml", package="simplaceUtil"))
getXMLParameterIDs(params, "CROPNAME")

}
