% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_to_graph.R
\name{graphLegend}
\alias{graphLegend}
\title{Creates legends}
\usage{
graphLegend(showinterfaces = TRUE, fillcolors = NULL, shapes = NULL)
}
\arguments{
\item{showinterfaces}{if TRUE, include interfaces in graph}

\item{fillcolors}{vector of colors, named by component's subtype}

\item{shapes}{vector of shapes, named by component's type}
}
\value{
a list with the component legend and links (arrows) legend
}
\description{
Creates legends
}
