% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{isSolution}
\alias{isSolution}
\title{Determines the solution by it's file extension}
\usage{
isSolution(file)
}
\arguments{
\item{file}{filename}
}
\description{
Determines the solution by it's file extension
}
