% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{mark_variables}
\alias{mark_variables}
\title{Add the prefix "variables" if there is none}
\usage{
mark_variables(v)
}
\arguments{
\item{v}{vector with prefix and name}
}
\description{
Add the prefix "variables" if there is none
}
