% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{markVariables}
\alias{markVariables}
\title{Add the prefix "variables" if there is none}
\usage{
markVariables(v)
}
\arguments{
\item{v}{vector with prefix and name}
}
\description{
Add the prefix "variables" if there is none
}
\keyword{internal}
