% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvoutput_helper.R
\name{parseDate}
\alias{parseDate}
\title{Converts date (class \code{character}) column from output file to column of class \code{Date}}
\usage{
parseDate(
  data,
  newName = "CURRENT.DATE",
  format = "\%d.\%m.\%Y",
  oldName = "CURRENT.DATE"
)
}
\arguments{
\item{data}{data.frame}

\item{newName}{name of the transformed date column}

\item{format}{date format}

\item{oldName}{name of the column that holds the date to be transformed}
}
\value{
data.frame with date column of class \code{Date}
}
\description{
Converts date (class \code{character}) column from output file to column of class \code{Date}
}
