% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotScalarOutput}
\alias{plotScalarOutput}
\title{Plots scalar output}
\usage{
plotScalarOutput(
  data,
  column_x,
  columns_y,
  simulationid = NULL,
  date_from = NULL,
  date_to = NULL,
  datecol = "CURRENT.DATE",
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{data}{data from memory output or file}

\item{column_x}{column name for x values}

\item{columns_y}{column name(s) for y values (vector of names)}

\item{simulationid}{plot only data for simulationids}

\item{date_from}{simulation date from where values are plotted}

\item{date_to}{simulateon date until values are plotted}

\item{datecol}{column name for date (default CURRENT.DATE)}

\item{nrow}{number of panel rows when plotting multiple simulation ids}

\item{ncol}{number of panels in a row when plotting multiple simulation ids}
}
\description{
Plots scalar output
}
