% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{readXMLParameterFile}
\alias{readXMLParameterFile}
\title{Read XML parameter file}
\usage{
readXMLParameterFile(file)
}
\arguments{
\item{file}{filename of parameterfile}
}
\value{
parameterdata (XML2 object)
}
\description{
Read XML parameter file
}
