% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{removeComponentInput}
\alias{removeComponentInput}
\title{Removes an input for component}
\usage{
removeComponentInput(sol, componentid, id)
}
\arguments{
\item{sol}{solution object}

\item{componentid}{id of the sim component}

\item{id}{id of the input}
}
\value{
modified solution object
}
\description{
Removes an input for component
}
