% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{removeOutput}
\alias{removeOutput}
\title{Removes output with given id}
\usage{
removeOutput(sol, outputid)
}
\arguments{
\item{sol}{solution object}

\item{outputid}{id of output to remove}
}
\value{
modified solution object
}
\description{
Notice: the interface for the output will also be removed.
}
