% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{removeOutputVariable}
\alias{removeOutputVariable}
\title{Removes an output variable from a given output}
\usage{
removeOutputVariable(sol, outputid, id)
}
\arguments{
\item{sol}{solution object}

\item{outputid}{id of output from where the variable should be removed}

\item{id}{name of the variable to remove}
}
\value{
modified solution object
}
\description{
Removes an output variable from a given output
}
