% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{replaceVariablesWithValues}
\alias{replaceVariablesWithValues}
\title{Replaces variables with content}
\usage{
replaceVariablesWithValues(text, variables, additional)
}
\arguments{
\item{text}{text to replace}

\item{variables}{variables dataframe}

\item{additional}{additional variables as named vector c("var1"="value1", ...)}
}
\description{
Replaces variables with content
}
