% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvoutput_helper.R
\name{resultToDataframeExpanded}
\alias{resultToDataframeExpanded}
\title{converts expanded resultlist to dataframe in wide format}
\usage{
resultToDataframeExpanded(reslist)
}
\arguments{
\item{reslist}{result list from simplace simulation}
}
\value{
data.frame with layered values (wide format)
}
\description{
converts expanded resultlist to dataframe in wide format
}
