% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{scaleXMLParameter}
\alias{scaleXMLParameter}
\title{Scales a parameter value by a value}
\usage{
scaleXMLParameter(paramdata, key, id, factor = 1)
}
\arguments{
\item{paramdata}{parameterdata (xml2 object)}

\item{key}{named vector \code{c(idname="value")} to identify the dataset, the name is the id of the parameter and the value has to match the parameter's value}

\item{id}{name of the parameter which should be scaled}

\item{factor}{scaling factor}
}
\value{
modified parameter set (xml2 object)
}
\description{
Scales a parameter value by a value
}
\examples{

params <- readXMLParameterFile(system.file("input", "crop.xml", package="simplaceUtil"))
getXMLParameter(params, c(CROPNAME="soy bean"), "RUETableRUE")
params_new <-scaleXMLParameter(params, c(CROPNAME="soy bean"), "RUETableRUE",0.9)
getXMLParameter(params_new, c(CROPNAME="soy bean"), "RUETableRUE")

}
