% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{setInputValue}
\alias{setInputValue}
\title{Sets the value of an input for a given parent element}
\usage{
setInputValue(sol, parentid, id, value, datatype = NULL)
}
\arguments{
\item{sol}{solution object}

\item{parentid}{id of input parent}

\item{id}{id of the input}

\item{value}{new value}

\item{datatype}{datatype (optional)}
}
\value{
modified solution object
}
\description{
Sets the value of an input for a given parent element
}
