% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{setInputValueForCategory}
\alias{setInputValueForCategory}
\title{Sets the value of an input in all elements of the given category}
\usage{
setInputValueForCategory(sol, category, id, value, datatype = NULL)
}
\arguments{
\item{sol}{solution object}

\item{category}{tag of categories containing inputs, e.g. \code{transform} or \code{simcomponent}}

\item{id}{id of the input}

\item{value}{new value}

\item{datatype}{datatype (optional)}
}
\value{
modified solution object
}
\description{
It's main use is to change consistently inputs of all transformers,
e.g. \code{layerthickness}. It can also be used to set sim component inputs
to a fixed value.
}
