% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{setXMLParameter}
\alias{setXMLParameter}
\title{Set a parameter value for a specific parameterset}
\usage{
setXMLParameter(paramdata, key, id, value = "")
}
\arguments{
\item{paramdata}{parameterdata (xml2 object)}

\item{key}{named vector \code{c(idname="value")} to identify the dataset, the name is the id of the parameter and the value has to match the parameter's value}

\item{id}{name of the parameter of which the value should be set}

\item{value}{new value to set (vector)}
}
\value{
modified parameter set (xml2 object)
}
\description{
Set a parameter value for a specific parameterset
}
\examples{

params <- readXMLParameterFile(system.file("input", "crop.xml", package="simplaceUtil"))
getXMLParameter(params, c(CROPNAME="soy bean"), "TSUM1")
params_new <-setXMLParameter(params, c(CROPNAME="soy bean"), "TSUM1",400)
getXMLParameter(params_new, c(CROPNAME="soy bean"), "TSUM1")

}
