% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionToGraphFunctions.R
\name{solutionToGraph}
\alias{solutionToGraph}
\title{Creates a graph from a solution}
\usage{
solutionToGraph(file, showinterfaces = FALSE)
}
\arguments{
\item{file}{solution file}

\item{showinterfaces}{if true include interfaces}
}
\description{
Creates a graph from a solution
}
