% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{subsetDataForPlot}
\alias{subsetDataForPlot}
\title{Subsets data for plotting}
\usage{
subsetDataForPlot(
  data,
  simulationid = NULL,
  date_from = NULL,
  date_to = NULL,
  datecol = "CURRENT.DATE"
)
}
\arguments{
\item{data}{data from memory output or file}

\item{simulationid}{get data only for simulationid}

\item{date_from}{simulation date from where values are taken}

\item{date_to}{simulateon date until values are taken}

\item{datecol}{column name for date (default CURRENT.DATA)}
}
\description{
Subsets data for plotting
}
\keyword{internal}
