% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{swapComponents}
\alias{swapComponents}
\title{Swaps the order of SimComponents in the solution}
\usage{
swapComponents(sol, order)
}
\arguments{
\item{sol}{solution object}

\item{order}{a vector of component positions}
}
\value{
modified solution object
}
\description{
Rearranges the order of SimComponents. All components that are mentioned in
the vector order will be rearranged according to their position in the order
vector. All other components remain on the same position.
E. g. an order of \code{c(5,6,1,3)} will put components on position 1,3,5,6 to
position 5, 6, 1, 3.
}
