% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvoutput_helper.R
\name{transformLayeredData}
\alias{transformLayeredData}
\title{Transform layered output data in long format}
\usage{
transformLayeredData(data, sep = "_")
}
\arguments{
\item{data}{dataframe}

\item{sep}{character that separates layer number from variable name (default "_")}
}
\value{
dataframe in long format
}
\description{
Transform layered output data in long format
}
