% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unitsymbols}
\alias{unitsymbols}
\title{Symbols for unit URIs}
\format{
A named vector of unit symbols where the names ar the unit URIs
}
\usage{
data(unitsymbols)
}
\description{
Symbols for unit URIs
}
\examples{
# data(unitsymbols)
# unitsymbols['http://www.wurvoc.org/vocabularies/om-1.8/gram_per_square_metre']

}
\keyword{datasets}
