% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{writeSolutionToFile}
\alias{writeSolutionToFile}
\title{Writes solution to file}
\usage{
writeSolutionToFile(sol, file)
}
\arguments{
\item{sol}{solution object (xml_document)}

\item{file}{filename for the solution}
}
\description{
Writes solution to file
}
