% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_handling.R
\name{writeXMLParameterFile}
\alias{writeXMLParameterFile}
\title{Write parameterdata to xml file}
\usage{
writeXMLParameterFile(paramdata, file)
}
\arguments{
\item{paramdata}{parameterdata (xml2 object)}

\item{file}{filename of parameterfile}
}
\value{
just for side effects, returs paramdata invisibly
}
\description{
Write parameterdata to xml file
}
