% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolutionToGraphFunctions.R
\name{componentsToGraph}
\alias{componentsToGraph}
\title{Creates a graph from component and links dataframe}
\usage{
componentsToGraph(comp, links)
}
\arguments{
\item{comp}{components dataframe}

\item{links}{links dataframe}
}
\description{
Creates a graph from component and links dataframe
}
