% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_to_graph.R
\name{componentsToGraph}
\alias{componentsToGraph}
\title{Creates a graph from component and links dataframe}
\usage{
componentsToGraph(comp, links, showinterfaces = FALSE)
}
\arguments{
\item{comp}{components dataframe}

\item{links}{links dataframe}

\item{showinterfaces}{if TRUE, include interfaces in graph}
}
\description{
Creates a graph from component and links dataframe
}
