% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_to_graph.R
\name{filterEdges}
\alias{filterEdges}
\title{Selects edges of specific timestep}
\usage{
filterEdges(graph, linkage = "allsteps")
}
\arguments{
\item{graph}{graph object of class \code{dgr_graph}}

\item{linkage}{all steps, same step or previous step}
}
\value{
graph object of class \code{dgr_graph}
}
\description{
Selects edges of specific timestep
}
