% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getComponents}
\alias{getComponents}
\title{Gets components for a solution}
\usage{
getComponents(x)
}
\arguments{
\item{x}{xml object (solution)}
}
\value{
data.frame with the solution components (resources, sim components, outputs)
}
\description{
Gets components for a solution
}
