% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getElementsFromSolutionFile}
\alias{getElementsFromSolutionFile}
\title{Get component and links dataframe from solution file}
\usage{
getElementsFromSolutionFile(file)
}
\arguments{
\item{file}{solution}
}
\value{
list with solution (xml2 object) and components, links and variables data.frame
}
\description{
Get component and links dataframe from solution file
}
