% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_to_graph.R
\name{getLinkingToComponent}
\alias{getLinkingToComponent}
\title{Get components that link to given components}
\usage{
getLinkingToComponent(
  graph,
  names,
  distance = 50,
  linkage = "allsteps",
  type = "all"
)
}
\arguments{
\item{graph}{graph object of class \code{dgr_graph}}

\item{names}{names (id attributes) of selected components}

\item{distance}{maximum distance from given components}

\item{linkage}{all steps, same step or previous step}

\item{type}{type of components}
}
\value{
graph object of class \code{dgr_graph}
}
\description{
Get components that link to given components
}
