% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_to_graph.R
\name{getNeighborhood}
\alias{getNeighborhood}
\title{Get the simcomponent in the neighborhood of given component}
\usage{
getNeighborhood(graph, name, distance = 1, linkage = "allsteps", type = "all")
}
\arguments{
\item{graph}{the solution graph}

\item{name}{id of the sim component}

\item{distance}{maximum number of steps from given component}

\item{linkage}{all steps, same step or previous step}

\item{type}{type of component}
}
\description{
Get the simcomponent in the neighborhood of given component
}
