% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getSolutionInfoAsDataframe}
\alias{getSolutionInfoAsDataframe}
\title{Get info for a solution}
\usage{
getSolutionInfoAsDataframe(file, workdir)
}
\arguments{
\item{file}{solution file}

\item{workdir}{working directory}
}
\description{
Get info for a solution
}
