% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getUserVariables}
\alias{getUserVariables}
\title{Gets user variables for a solution}
\usage{
getUserVariables(x)
}
\arguments{
\item{x}{xml object (solution)}
}
\value{
data.frame with the user defined variable
}
\description{
Gets user variables for a solution
}
