% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{getVars}
\alias{getVars}
\title{Get the variables from source, rule etc. attributes}
\usage{
getVars(s, ruleonly = FALSE)
}
\arguments{
\item{s}{attribute text (rule)}

\item{ruleonly}{check only for ${} variables}
}
\description{
Get the variables from source, rule etc. attributes
}
\keyword{internal}
