% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvoutput_helper.R
\name{parseDate}
\alias{parseDate}
\title{Converts date from output file to date object}
\usage{
parseDate(
  data,
  newName = "CURRENT.DATE",
  format = "\%d.\%m.\%Y",
  oldName = "CURRENT.DATE"
)
}
\arguments{
\item{data}{dataframe}

\item{newName}{name of the transformed date column}

\item{format}{date format}

\item{oldName}{name of the column that holds the date to be transformed}
}
\description{
Converts date from output file to date object
}
