% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLayeredOutput}
\alias{plotLayeredOutput}
\title{Plots layered output}
\usage{
plotLayeredOutput(
  data,
  column,
  simulationid = NULL,
  date_from = NULL,
  date_to = NULL,
  datecol = "CURRENT.DATE"
)
}
\arguments{
\item{data}{data from memory output or file}

\item{column}{column name used for the fill color}

\item{simulationid}{plot only data for simulationid}

\item{date_from}{simulation date from where values are plotted}

\item{date_to}{simulateon date until values are plotted}

\item{datecol}{column name for date (default CURRENT.DATA)}
}
\description{
Plots layered output
}
