% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_modifications.R
\name{replaceVariable}
\alias{replaceVariable}
\title{Replaces a variable with another.}
\usage{
replaceVariable(sol, oldid, newid)
}
\arguments{
\item{sol}{solution object}

\item{oldid}{id of variable to be replaced}

\item{newid}{id of the replacing variable}
}
\value{
modified solution object
}
\description{
Notice: there is no check, whether the ids exist. Variables are replaced
literarilly when they are the only content of an attribute or element content.
In other attributes and element contents ${oldid} will be replaced by ${newid}
}
