% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{runSimplaceCustomGuiApp}
\alias{runSimplaceCustomGuiApp}
\title{Runs Custom Shiny app}
\usage{
runSimplaceCustomGuiApp(
  simplacedirs,
  solution,
  project,
  paramlist,
  plotlist,
  ...
)
}
\arguments{
\item{simplacedirs}{a list containing the named elements instdir, workdir, outdir and options}

\item{solution}{filename of a solution (or solution as text)}

\item{project}{filename of a project}

\item{paramlist}{list of named parameter elements}

\item{plotlist}{list of named functions that produce a plot}

\item{...}{parameters passed to shiny::runApp}
}
\value{
does not return a value, called for the side effect of running the shiny app
}
\description{
User can create automatically a simple GUI for a specific solution/project
by specifying solution, project, list of parameters that could be interatively
changed as well a list of custom plot functions.
}
