% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions_elements.R
\name{simComponentType}
\alias{simComponentType}
\title{Determines the type of sim component by class attribute}
\usage{
simComponentType(s)
}
\arguments{
\item{s}{class attribute}
}
\description{
Determines the type of sim component by class attribute
}
