% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplaceUtil.R
\docType{package}
\name{simplaceUtil}
\alias{simplaceUtil-package}
\alias{simplaceUtil}
\title{Package with utility function for working with Simplace}
\description{
Provides functions to work with the modeling framework Simplace
}
\details{
\itemize{
\item get elements of a simplace solution as dataframes
\item visualise the structure of a solution as a graph
\item helper functions for transforming output data
\item simplified plot functions for simulation output variables
\item shinyApp to visualise graphs, run simulations and plot results
}
}
\examples{
\dontrun{
# run the GUI
runSimplaceGuiApp()
}

# visualise the solution structure as graph
graph <- solutionToGraph(system.file("solution", "Yield.sol.xml",
  package = "simplaceUtil"))
DiagrammeR::render_graph(graph)
}
\author{
{Gunther Krauss}
}
