% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvoutput_helper.R
\name{transformLayeredData}
\alias{transformLayeredData}
\title{Transform layered output data in long format}
\usage{
transformLayeredData(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Transform layered output data in long format
}
